{{ cookiecutter.repo_name }}/requirements.txt
click
Sphinx
coverage
awscli
flake8
python-dotenv
numpy
pandas
matplotlib
"beautifulsoup4"